move AllowSiblingsOnly to core
