+++ parachain/pallets/control/src/lib.rs
@@ -64,7 +64,7 @@ use xcm_executor::traits::ConvertLocation;
 
 use frame_support::{
 	pallet_prelude::*,
-	traits::{tokens::Preservation, Contains, EnsureOrigin},
+	traits::{tokens::Preservation, EnsureOrigin},
 };
 use frame_system::pallet_prelude::*;
 use snowbridge_core::{
@@ -120,13 +120,6 @@ where
 	Partial(AccountIdOf<T>),
 	/// No charge
 	No,
-}
-
-pub struct AllowSiblingsOnly;
-impl Contains<MultiLocation> for AllowSiblingsOnly {
-	fn contains(location: &MultiLocation) -> bool {
-		matches!(location, MultiLocation { parents: 1, interior: X1(Parachain(_)) })
-	}
 }
 
 #[frame_support::pallet]
